pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "webnose"
version = "2.4.0"
description = "A modular web smell scanner"
readme = "README.md"
authors = [
  { name="mllamazares", email="your.email@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.7"
dependencies = [
    "requests",
    "PyYAML",
    "tqdm",
    "fake-useragent",
    "urllib3"
]

[project.scripts]
webnose = "webnose:main"

[tool.setuptools]
py-modules = ["webnose"]
